requirements.txt
pip
numpy
pandas
flask
gunicorn
scikit-learn
xgboost
python-dotenv